More build fixes
Canonical link: https://commits.webkit.org/35251@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@43227 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ JavaScriptCore/jit/JIT.cpp
@@ -2011,7 +2011,7 @@ void JIT::privateCompileCTIMachineTrampolines(RefPtr<ExecutablePool>* executable
     storePtr(regT0, Address(stackPointerRegister, FIELD_OFFSET(NativeFunctionSignature, argPointer)));
 
     // regT1 currently points to the first argument, regT1 - sizeof(Register) points to 'this'
-    loadPtr(Address(regT1, -sizeof(Register)), regT1);
+    loadPtr(Address(regT1, -(int)sizeof(Register)), regT1);
     poke(regT1, 2);
     storePtr(regT1, Address(stackPointerRegister, FIELD_OFFSET(NativeFunctionSignature, thisValue)));
